requirements.txt
functions-framework==3.*
google-play-scraper==1.2.4
google-cloud-pubsub>=2.10.0